ホーム画面のAPIとAction,Reducer追加

+++ src/app/api/login/route.ts
@@ -5,7 +5,7 @@ import { ServerHandler } from "@/lib/server/util/db_util"
 import { StringUtils } from "@/lib/common/util/string_utils";
 import { SessionDAO } from "@/lib/server/dao/session";
 import { UserAccountDAO } from "@/lib/server/dao/user_account";
-import { AuthenticationExeption } from "@/lib/server/util/exeption";
+import { AuthenticationExeption, ServerExeption } from "@/lib/server/util/exeption";
 
 export const dynamic = 'force-dynamic'
 
@@ -34,7 +34,11 @@ export async function POST(
             throw new AuthenticationExeption()
         }
 
-        newToken = await SessionDAO.insert(client, u.id as number)
+        if (u.id === undefined) {
+            throw new ServerExeption()
+        }
+
+        newToken = await SessionDAO.insert(client, u.id)
 
         await SessionDAO.deleteExpiredSession(client)
